get wallet created as common one
