+++ frontend/marketplace/src/components/CertificateTabContent.vue
@@ -8,7 +8,7 @@ import { toast } from "vue3-toastify";
 import { useQuery } from "@vue/apollo-composable";
 import gql from "graphql-tag";
 import CustomAlert from "@/components/CustomAlert.vue";
-import {walletConnected} from "@/utils/utils";
+import {walletConnected, getWallet} from "@/utils/wallet-utils";
 
 const pageNumber = ref(1)
 const itemsPerPage = ref(5)
@@ -26,7 +26,8 @@ const handleSearch = () => {
 
 const getCertificatesData = async () => {
   try {
-    const account = await window.keplr.getKey(CHAIN_ID);
+    const wallet = getWallet()
+    const account = await wallet.getKey(CHAIN_ID);
     const walletAddress = account.bech32Address
     if (walletAddress) {
       const query = `query {
